Outfactor annotation management functionality to annotatable.view.js
